+++ business/src/main/java/com/kevinguanchedarias/owgejava/configurations/QuartzConfiguration.java
@@ -1,9 +1,0 @@
-package com.kevinguanchedarias.owgejava.configurations;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.ImportResource;
-
-@Configuration
-@ImportResource("META-INF/quartz-context.xml")
-public class QuartzConfiguration {
-}