+++ game-rest/src/main/java/com/kevinguanchedarias/owgejava/OwgeRestApplication.java
@@ -3,12 +3,10 @@ package com.kevinguanchedarias.owgejava;
 import com.kevinguanchedarias.taggablecache.configuration.TaggableCacheDefaultConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.quartz.SchedulerFactoryBeanCustomizer;
 import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 import org.springframework.boot.autoconfigure.security.servlet.SecurityFilterAutoConfiguration;
 import org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration;
 import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.data.jdbc.repository.config.EnableJdbcRepositories;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
@@ -31,11 +29,6 @@ public class OwgeRestApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(OwgeRestApplication.class, args);
-    }
-
-    @Bean
-    public SchedulerFactoryBeanCustomizer schedulerFactoryBeanCustomizer() {
-        return bean -> bean.setTaskExecutor(null);
     }
 
 }